requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
base58==2.1.0
beautifulsoup4==4.10.0
bitarray==1.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
h11==0.12.0
h2==3.2.0
hexbytes==0.2.2
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipfshttpclient==0.8.0a2
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
jsons==1.5.1
jsonschema==3.2.0
lru-dict==1.1.7
lxml==4.6.3
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
outcome==1.1.0
parsel==1.6.0
parsimonious==0.8.1
priority==1.3.0
Protego==0.1.16
protobuf==3.19.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pycryptodomex==3.11.0
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
pyrsistent==0.18.0
queuelib==1.6.2
requests==2.26.0
rlp==2.0.1
rsa==4.7.2
Scrapy==2.5.1
selenium==4.0.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
telepota==1.0
Telethon==1.23.0
toolz==0.11.1
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
typing-extensions==3.10.0.2
typish==1.9.3
urllib3==1.26.7
varint==1.0.2
w3lib==1.22.0
web3==5.24.0
websockets==9.1
wsproto==1.0.0
yarl==1.7.0
zope.interface==5.4.0
